Fix an invalid argument error from sendto.
Add a lot of log messages to see what's happening. Currently,
the program is dominated by the Printer and MessageSender threads.
And it is not switching between them too often it appears either.
When I say dominated--well, the console prints so many messages
it's hard to tell, but it appears that Printer and MessageSender
are the only threads running--so something weird is happening.